# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: WG-Split API
  version: 1.0.0
servers:
  - url: https://api.wg-split.example.com
paths:
  /auth/magic-link/request:
    post:
      summary: Request a magic login link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MagicLinkRequest'
      responses:
        '200':
          description: Magic link issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MagicLinkResponse'
  /auth/magic-link/consume:
    post:
      summary: Consume a magic link token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MagicLinkConsume'
      responses:
        '200':
          description: JWT issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtResponse'
  /me:
    get:
      summary: Current user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User with memberships
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /groups:
    get:
      summary: List groups for current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Collection of groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Group'
    post:
      summary: Create group
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupCreate'
      responses:
        '201':
          description: Group created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
  /groups/{groupId}:
    get:
      summary: Get group detail
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '200':
          description: Group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupDetail'
    patch:
      summary: Update group settings
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupUpdate'
      responses:
        '200':
          description: Updated group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
    delete:
      summary: Delete group
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '204':
          description: Deleted
  /groups/{groupId}/expenses:
    get:
      summary: List expenses
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupId'
        - in: query
          name: from
          schema:
            type: string
            format: date
        - in: query
          name: to
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Expense'
    post:
      summary: Create expense
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseCreate'
      responses:
        '201':
          description: Expense created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    GroupId:
      in: path
      name: groupId
      required: true
      schema:
        type: string
  schemas:
    MagicLinkRequest:
      type: object
      properties:
        email:
          type: string
          format: email
      required:
        - email
    MagicLinkResponse:
      type: object
      properties:
        token:
          type: string
        expiresAt:
          type: string
          format: date-time
    MagicLinkConsume:
      type: object
      properties:
        token:
          type: string
      required:
        - token
    JwtResponse:
      type: object
      properties:
        jwt:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        memberships:
          type: array
          items:
            $ref: '#/components/schemas/GroupMember'
    Group:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        currency:
          type: string
        rounding:
          type: integer
    GroupDetail:
      allOf:
        - $ref: '#/components/schemas/Group'
        - type: object
          properties:
            members:
              type: array
              items:
                $ref: '#/components/schemas/GroupMember'
    GroupMember:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        role:
          type: string
    GroupCreate:
      type: object
      properties:
        name:
          type: string
        currency:
          type: string
        rounding:
          type: integer
      required:
        - name
        - currency
    GroupUpdate:
      type: object
      properties:
        name:
          type: string
        currency:
          type: string
        rounding:
          type: integer
    Expense:
      type: object
      properties:
        id:
          type: string
        amount:
          type: number
        currency:
          type: string
        category:
          type: string
        date:
          type: string
          format: date-time
    ExpenseCreate:
      type: object
      properties:
        payerId:
          type: string
        amount:
          type: number
        currency:
          type: string
        date:
          type: string
        category:
          type: string
        splitType:
          type: string
          enum: [EQUAL, WEIGHTED, SHARES, FIXED, PERCENT]
        participants:
          type: array
          items:
            type: object
            properties:
              userId:
                type: string
              share:
                type: number
              weight:
                type: number
              fixedAmount:
                type: number
              percent:
                type: number
            required:
              - userId
      required:
        - payerId
        - amount
        - category
        - splitType
        - participants